+++ projects/clr/hipamd/src/hip_executionctx.hpp
@@ -25,7 +25,7 @@ typedef struct DevResourceDesc {
 
 class ExecutionCtx {
 public:
-  ExecutionCtx(int deviceId, DevResourceDesc* desc, unsigned int flags);
+  ExecutionCtx(int deviceId, DevResourceDesc* desc, uint32_t flags);
   ~ExecutionCtx();
 
   hipError_t Create();
@@ -33,9 +33,9 @@ public:
   int deviceId() const { return deviceId_; }
   hip::Device* device() const { return g_devices[deviceId_]; }
   const std::vector<uint32_t>& cuMask() const { return cuMask_; }
-  unsigned int cuCount() const { return cuCount_; }
-  unsigned int flags() const { return flags_; }
-  unsigned long long ctxId() const { return ctxId_; }
+  uint32_t cuCount() const { return cuCount_; }
+  uint32_t flags() const { return flags_; }
+  uint64_t ctxId() const { return ctxId_; }
   std::recursive_mutex& lock() { return lock_; }
 
   void addStream(hip::Stream* stream);
@@ -47,16 +47,16 @@ public:
 
   static hipError_t deviceGetDevResource(int device, hipDevResource* resource,
                                          hipDevResourceType type);
-  static hipError_t devSmResourceSplitByCount(hipDevResource* result, unsigned int* nbGroups,
+  static hipError_t devSmResourceSplitByCount(hipDevResource* result, uint32_t* nbGroups,
                                               const hipDevResource* input,
                                               hipDevResource* remainder,
-                                              unsigned int flags, unsigned int minCount);
-  static hipError_t devSmResourceSplit(hipDevResource* result, unsigned int nbGroups,
+                                              uint32_t flags, uint32_t minCount);
+  static hipError_t devSmResourceSplit(hipDevResource* result, uint32_t nbGroups,
                                        const hipDevResource* input, hipDevResource* remainder,
-                                       unsigned int flags,
+                                       uint32_t flags,
                                        hipDevSmResourceGroupParams* groupParams);
   static hipError_t devResourceGenerateDesc(hipDevResourceDesc_t* phDesc,
-                                            hipDevResource* resources, unsigned int nbResources);
+                                            hipDevResource* resources, uint32_t nbResources);
 
   static ExecutionCtx* createPrimaryCtx(int device);
 
@@ -67,9 +67,9 @@ private:
   ExecutionCtx& operator=(ExecutionCtx&&) = delete;
 
   int deviceId_;
-  unsigned int flags_;
-  unsigned int cuCount_;
-  unsigned long long ctxId_;
+  uint32_t flags_;
+  uint32_t cuCount_;
+  uint64_t ctxId_;
   std::vector<uint32_t> cuMask_;
   DevResourceDesc* resourceDesc_;
 
@@ -77,7 +77,7 @@ private:
   std::shared_mutex streamSetLock_;
   std::unordered_set<hip::Stream*> streams_;
 
-  inline static std::atomic<unsigned long long> nextCtxId_{1};
+  inline static std::atomic<uint64_t> nextCtxId_{1};
   inline static std::atomic<uint32_t> nextResourceId_{1};
   inline static std::atomic<uint32_t> nextFamilyId_{1};
 
@@ -85,15 +85,15 @@ private:
   static uint32_t readResourceId(const hipDevResource* res);
   static int readDeviceId(const hipDevResource* res);
   static void registerResourceMeta(int deviceId, uint32_t resourceId,
-                                   uint32_t familyId, unsigned int startCU);
+                                   uint32_t familyId, uint32_t startCU);
   static const ResourceMeta* lookupResourceMeta(int deviceId, uint32_t resourceId);
 
-  static void fillSmResult(hipDevResource* res, unsigned int smCount,
-                           unsigned int alignment, unsigned int flags);
-  static void fillRemainder(hipDevResource* remainder, unsigned int remainingCUs,
-                            unsigned int alignment);
-  static std::vector<uint32_t> buildCuMask(unsigned int startCU, unsigned int count,
-                                           unsigned int totalCUs);
+  static void fillSmResult(hipDevResource* res, uint32_t smCount,
+                           uint32_t alignment, uint32_t flags);
+  static void fillRemainder(hipDevResource* remainder, uint32_t remainingCUs,
+                            uint32_t alignment);
+  static std::vector<uint32_t> buildCuMask(uint32_t startCU, uint32_t count,
+                                           uint32_t totalCUs);
 };
 
 } // namespace hip
